empty commit with explanation
I changed the README which used to say the months were an affine space.  The
months are not an affine space with integers as vectors between them because
the integers don't form a vector space. While the usual integers are a ring
with addition and multiplication, they are not a field which also requires
multiplicative inverses for the non zero elements. Vector spaces require that
their scalars are from a field. I tried to find an equivalent formulation of
affine spaces but for "vectors spaces over rings" (known as modules) but I
couldn't find anything good. It turns out that the correct structure is a
principal homogeneous space for a group.  In retrospect the ring structure's
multiplication is irrelevant to the situation.